Issue #1791: Aggregate Row Layout
Break out RowLayout into a separate file.
